+++ geotrek/settings/base.py
@@ -438,19 +438,24 @@ LAND_BBOX_CITIES_ENABLED = True
 LAND_BBOX_DISTRICTS_ENABLED = True
 LAND_BBOX_AREAS_ENABLED = False
 
+PUBLISHED_BY_LANG = True
+EXPORT_MAP_IMAGE_SIZE = {
+    'trek': (14.1, 11),
+    'poi':  (14.1, 11),
+}
+COMPLETENESS_FIELDS = {
+    'trek': ['departure', 'duration', 'difficulty', 'description_teaser']
+}
+
 TRAIL_MODEL_ENABLED = True
 TREKKING_TOPOLOGY_ENABLED = True
 TREK_POI_INTERSECTION_MARGIN = 500  # meters (used only if TREKKING_TOPOLOGY_ENABLED = False)
 
-TREK_PUBLISHED_BY_LANG = True
 TREK_POINTS_OF_REFERENCE_ENABLED = True
 TREK_EXPORT_POI_LIST_LIMIT = 14
 TREK_EXPORT_INFORMATION_DESK_LIST_LIMIT = 2
-TREK_EXPORT_MAP_IMAGE_SIZE = (14.1, 11)
 TREK_EXPORT_HEADER_IMAGE_SIZE = (10.7, 5.035)  # Keep ratio of THUMBNAIL_ALIASES['print']
 
-TREK_COMPLETENESS_FIELDS = ['departure', 'duration', 'difficulty',
-                            'description_teaser']
 TREK_DAY_DURATION = 10  # Max duration to be done in one day
 TREK_ICON_SIZE_POI = 18
 TREK_ICON_SIZE_PARKING = 18